pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fotoxi"
version = "0.1.0"
description = "Photo management app with FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.21",
    "exifread>=3.0",
    "imagehash>=4.3",
    "Pillow>=11.0",
    "httpx>=0.28",
    "python-multipart>=0.0.18",
    "pillow-heif>=0.18",
    "opencv-python-headless>=4.10",
    "alembic>=1.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "pytest-httpx>=0.35",
    "piexif",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
